add readme texts
