+++ SYSC4806_ShopifyProject-m1-dev-isaiah/README.md
@@ -1,1 +1,11 @@
 # SYSC4806_ShopifyProject
+
+Introduction:
+Welcome to our Mini Shopify Web Application! This web application provides a platform for users to create own shops, choose products to sell and
+allow customers to buy from your shops
+
+Technologies:
+Spring MVC framework.
+
+Launch:
+Please run the application MiniShopify, and once the application finishes its launch, users can link to localhost:8080 to interact with the interface
