+++ SYSC4806_ShopifyProject-m1-dev-isaiah/src/main/java/org/example/Customer.java
@@ -1,5 +1,63 @@
-package org.example;
 
+package org.example;
+import jakarta.persistence.*;
+@Entity
 public class Customer {
+
+    @Id
+    @GeneratedValue
+    private Long id = null;
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String phone;
+
+    @ManyToOne
+    private Shop shop;
+
+    public Customer(String firstName, String lastName, String email, String phone) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.phone = phone;
+    }
+
+    // Default constructor (satisfy JPA)
+    public Customer() {}
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+
+
 
 }
