--- src/main/java/org/example/Category.java
@@ -1,9 +1,0 @@
-package org.example;
-
-public enum Category {
-    CLOTHES,
-    APPLIANCES,
-    JEWELRY,
-    DECOR,
-    ELECTRONICS
-}
