--- src/main/java/org/example/Merchant.java
@@ -1,4 +1,0 @@
-package org.example;
-
-public class Merchant {
-}
