--- src/main/java/org/example/ProductRepository.java
@@ -1,6 +1,0 @@
-package org.example;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ProductRepository extends CrudRepository<Product, Long> {
-}
