--- src/main/java/org/example/ShopRepository.java
@@ -1,6 +1,0 @@
-package org.example;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ShopRepository extends CrudRepository<Shop, Long> {
-}
